LayoutTests/imported/w3c/web-platform-tests/css/geometry/DOMRectList-expected.txt

x

PASS DOMRectList is exposed
PASS DOMRectList is not [LegacyArrayClass]
PASS DOMRectList length
PASS DOMRectList indexed getter
PASS DOMRectList item()
